Extract voice activity detection logic to the separate file
